requirements.txt
sqlalchemy>=1.4.46
pyrogram>=2.0.57
tgcrypto>=1.2.3
Babel>=2.10.3
bcrypt>=3.2.2
PyJWT>=2.6.0